Fix for the time to convergence measure.
New approach to calculate time to convergence also takes into account
the edge case when the run 'leaves' gmp +- tolerance area in
a following subrun again. If that is the case, the run
was considered as converged before, now the convergence time
is either increased or set to None, as intended.